refactor: Controller interface 상위 계층으로 추가
