+++ tomcat/src/main/java/nextstep/jwp/controller/Controller.java
@@ -1,0 +1,8 @@
+package nextstep.jwp.controller;
+
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
+
+public interface Controller {
+    HttpResponse getResponse(HttpRequest httpRequest) throws Exception;
+}
