+++ tomcat/src/main/java/nextstep/jwp/controller/ControllerAdvice.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.controller;
+package nextstep.jwp.controller;
 
 import java.io.IOException;
 import org.apache.coyote.http11.HttpStatus;
