+++ tomcat/src/main/java/nextstep/jwp/controller/ControllerMapper.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.controller;
+package nextstep.jwp.controller;
 
 import java.util.Arrays;
 import java.util.function.Function;
@@ -7,18 +7,17 @@ import java.util.function.Predicate;
 public enum ControllerMapper {
     LOGIN(url -> url.contains("/login"), controller -> new LoginController()),
     REGISTER(url -> url.contains("/register"), controller -> new RegisterController()),
-    HOME(url -> url.contains("/"), controller -> new HomePageController())
-    ;
+    HOME(url -> url.contains("/"), controller -> new HomePageController());
 
     private final Predicate<String> findUrl;
-    private final Function<String, Controller> findController;
+    private final Function<String, AbstractController> findController;
 
-    ControllerMapper(Predicate<String> findUrl, Function<String, Controller> findController) {
+    ControllerMapper(Predicate<String> findUrl, Function<String, AbstractController> findController) {
         this.findUrl = findUrl;
         this.findController = findController;
     }
 
-    public static Controller findController(String requestUri) {
+    public static AbstractController findController(String requestUri) {
         return findUrl(requestUri)
                 .findController.apply(requestUri);
     }
