+++ tomcat/src/main/java/nextstep/jwp/controller/SessionManager.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.controller;
+package nextstep.jwp.controller;
 
 import java.util.HashMap;
 import java.util.Map;
